refactor send message page
