+++ website/frontend/src/components/common/Form/Form.module.scss
@@ -1,0 +1,79 @@
+@use '@gear-js/ui/variables' as *;
+@use '@gear-js/ui/breakpoints' as *;
+
+.largeForm {
+  display: flex;
+  flex-direction: column;
+  width: 80%;
+  
+  @include lg {
+    width: 100%;
+  }
+}
+
+.formItem {
+  position: relative;
+  margin-bottom: 25px;
+}
+
+.field {
+  display: flex;
+  justify-content: space-between;
+
+  & > span {
+    flex: 1 0 180px;
+    line-height: 1.3;
+    margin-right: 20px;
+    color: $textColorSecondary;
+    font-size: $fontSizeMedium;
+
+    &::after {
+      content: ':';
+    }
+    
+    @include md {
+      margin-bottom: 4px;
+      margin-right: 0;
+    }
+  }
+
+  & > span + * {
+    flex: 1 1 100%;
+  }
+
+  @include md {
+    align-items: stretch;
+    flex-direction: column;
+    justify-content: flex-start;
+
+    & > span,
+    & > span + * {
+      flex: none;
+    }
+  }
+}
+
+.formButtons {
+  display: flex;
+  margin-top: 5px;
+  
+  & > *:not(:last-child){
+    margin-right: 24px;
+
+    @include md {
+      margin-right: 0;
+      margin-bottom: 24px;
+    }
+  }
+
+  @include md {
+    flex-direction: column;
+  }
+}
+
+.error {
+  position: absolute;
+  color: $errorColor;
+  font-size: $fontSizeMedium;
+  bottom: -20px;
+}
