+++ website/frontend/src/components/common/Form/FormTextarea/FormTextarea.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 import { useField } from 'formik';
 import { Textarea, TextareaProps } from '@gear-js/ui';
 
-import styles from '../FormFields.module.scss';
+import styles from '../Form.module.scss';
 
 type Props = TextareaProps & {
   name: string;
@@ -16,7 +16,7 @@ const FormTextarea = (props: Props) => {
   const showError = meta.error && meta.touched;
 
   return (
-    <div className={styles.item}>
+    <div className={styles.formItem}>
       <Textarea {...other} {...field} label={label} className={clsx(styles.field, className)} />
       {showError && <div className={styles.error}>{meta.error}</div>}
     </div>
