+++ website/frontend/src/components/common/Form/index.ts
@@ -1,3 +1,6 @@
+import formStyles from './Form.module.scss';
+
 export * from './FormInput';
 export * from './FormTextarea';
 export * from './FormNumberFormat';
+export { formStyles };
