--- website/frontend/src/components/common/FormFields/FormFields.module.scss
@@ -1,32 +1,0 @@
-@use '@gear-js/ui/variables' as *;
-@use '@gear-js/ui/breakpoints' as *;
-
-.item {
-  position: relative;
-  margin-bottom: 25px;
-}
-
-.field {
-  @include md {
-    align-items: stretch;
-    flex-direction: column;
-  }
-
-  > span {
-    font-size: $fontSizeMedium;
-    line-height: 1.3;
-
-    @include md {
-      margin-bottom: 4px;
-      margin-right: 0;
-      min-width: none;
-    }
-  }
-}
-
-.error {
-  position: absolute;
-  color: $errorColor;
-  font-size: $fontSizeMedium;
-  bottom: -20px;
-}
