--- website/frontend/src/components/pages/Send/Send.scss
@@ -1,9 +1,0 @@
-@use 'assets/scss/variables' as *;
-
-.send-message__block {
-  background-color: $bkgActiveButtonColor;
-  border-radius: 2px;
-  border: 1px dashed #0e0e0e;
-  margin: 0 0 40px;
-  padding: 30px 25px;
-}
