--- website/frontend/src/components/pages/Send/children/MessageForm/MessageForm.scss
@@ -1,205 +1,0 @@
-@use 'assets/scss/variables' as *;
-
-.message-form__button {
-  background: #2bd071;
-  box-shadow: 0 0 24px rgba(43, 208, 113, 0.12);
-  border-radius: 2px;
-  color: #fff;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  gap: 17px;
-  padding: 12px;
-  font-weight: bold;
-  font-size: 16px;
-  line-height: 24px;
-  width: 194px;
-  margin-top: 32px;
-
-  &.disabled,
-  &[disabled] {
-    background: #5bffa5;
-    cursor: not-allowed;
-  }
-}
-
-.message-form__input-error {
-  background: #141010 !important;
-  border: 1px solid #8b3939 !important;
-}
-
-.message-form__error {
-  position: absolute;
-  color: #995757;
-  font-weight: normal;
-  font-size: 12px;
-  line-height: 130%;
-  bottom: -18px;
-}
-
-.message-form__field-wrapper {
-  width: 100%;
-  position: relative;
-}
-
-.message-form--switch {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-bottom: 40px;
-}
-
-.message-form--switch:first-child {
-  margin-top: 10px;
-}
-
-.message-form--switch:last-child {
-  margin-bottom: 0;
-}
-
-.message-form--switch__name {
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 24px;
-  color: #ffffff;
-  min-width: 100px;
-  text-align: right;
-}
-
-.message-form form {
-  background: #212121;
-  border: 1px dashed #0e0e0e;
-  border-radius: 2px;
-  padding: 30px 24px;
-  margin: 40px;
-}
-
-.message-form--wrapper {
-  display: flex;
-  flex-direction: row;
-}
-
-.message-form--col {
-  display: flex;
-  flex-direction: column;
-}
-
-.message-form--col:first-child {
-  margin-right: 24px;
-  width: 85%;
-}
-
-.message-form--info {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 24px;
-  width: 100%;
-}
-
-.message-form--info:last-child {
-  margin-bottom: 0;
-}
-
-.message-form--info .message-form__field {
-  color: #858585;
-  font-weight: 300;
-  font-size: 21px;
-  line-height: 1.85;
-  min-width: 164px;
-}
-
-.message-form--info .inputField {
-  height: 38px;
-  width: 100%;
-}
-
-.message-form--info .inputField {
-  padding: 10px 16px;
-  background: #000000;
-  border: none;
-  outline: none;
-  color: #fff;
-  width: 100%;
-  box-sizing: border-box;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 21px;
-  min-height: 38px;
-}
-
-.message-form--info .inputField::placeholder {
-  color: #8b8b8b;
-}
-
-.message-form--payload p {
-  color: #858585;
-  font-weight: 600;
-  font-size: 21px;
-  line-height: 130%;
-  min-width: 164px;
-  margin-bottom: 24px;
-}
-
-.message-form--btns {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 24px;
-}
-
-@media (max-width: 768px) {
-  .message-form--btns {
-    flex-direction: column;
-  }
-  .message-form__button {
-    width: 100%;
-  }
-
-  .message-form__button img {
-    display: none;
-  }
-
-  .message-form--info {
-    flex-direction: column;
-    align-items: unset;
-    margin-bottom: 32px;
-  }
-
-  .message-form--info label {
-    margin-bottom: 4px;
-  }
-
-  .message-form--info:first-child {
-    display: flex;
-    flex-direction: column;
-  }
-
-  .message-form--info:first-child span:last-child {
-    font-weight: normal;
-    font-size: 21px;
-    line-height: 130%;
-    color: #e0e0e0;
-  }
-
-  .message-form--info:first-child span:first-child {
-    color: #858585;
-    font-weight: 300;
-    font-size: 21px;
-    line-height: 130%;
-    min-width: 164px;
-  }
-
-  .message-form form {
-    background: none;
-    border: none;
-    border-radius: 0;
-    padding: 27px 16px;
-    margin: 0;
-  }
-
-  .message-form--col:first-child {
-    margin-right: 0;
-    width: 100%;
-  }
-}
