+++ website/frontend/src/components/pages/Send/children/MessageForm/children/PayloadType/PayloadType.module.scss
@@ -2,7 +2,6 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  width: 100%;
 }
 
 .checkbox {
