--- website/frontend/src/components/pages/Send/children/MessageForm/styles.ts
@@ -1,6 +1,0 @@
-import { styled } from 'stitches.config';
-
-export const MetaErrorMessage = styled('div', {
-  marginBottom: '16px',
-  color: '$white',
-});
